organized files & added comments/implemented remaining non-gfx funcs * created comments or implemented all non-graphics api related functions * organized fils by doxygen defgroup in RFGW.h; * separated consts into their own files for easier readability * renamed helpers for readability * added windows registry (see internal_window.go) which CreateWindow and Close now use * removed files that did not represent doxygen defgroups

+++ rgfw/internal_helps.go
@@ -1,6 +1,6 @@
 package rgfw
 
-func boolToInt(b bool) int {
+func boolToInt(b bool) uint8 {
     if b {
         return 1
     }
